requirements.txt
awscli==1.11.188
boto3==1.4.7
Pillow==4.3.0
requests==2.18.4
